requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
decorator==4.4.2
Flask==2.3.2
idna==3.4
imageio==2.30.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
moviepy==1.0.3
numpy==1.24.3
Pillow==9.5.0
proglog==0.1.10
requests==2.31.0
tqdm==4.65.0
urllib3==2.0.2
Werkzeug==2.3.4
zipp==3.15.0
